[TAN-1664] Move to new ProjectBasketsGenerator service
