+++ back/app/controllers/web_api/v1/projects_controller.rb
@@ -146,7 +146,7 @@ class WebApi::V1::ProjectsController < ApplicationController
   def voting_xlsx
     if @project.phases.where(participation_method: 'voting').present?
       I18n.with_locale(current_user&.locale) do
-        xlsx = XlsxService.new.generate_project_voting_xlsx @project.id
+        xlsx = XlsxExport::ProjectBasketsGenerator.new(@project).generate_project_baskets_xlsx
         send_data xlsx, type: 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet',
           filename: 'voter_selections.xlsx'
       end
